Add sync to persistant

+++ src/graph/mod.rs
@@ -48,10 +48,10 @@ pub struct Nodes {
 pub struct Node {
     pub id: String,
     pub label: String,
-    pub x: u32,
-    pub y: u32,
-    pub width: u32,
-    pub height: u32,
+    pub x: f32,
+    pub y: f32,
+    pub width: f32,
+    pub height: f32,
     pub background: String,
 }
